Updates models and correct the error in ptb-lm
It is not using the initialization they are asking for currently. That one makes the gradient explode...
Still, need to test to see if I can get to the value they said we should get,
For the GRU there is 2 versions the normal one and the faster one. 
The faster one is slightly faster still 2 times slower than the nn.GRU. 
Also It uses the same formula as the GRU from nn. and not the one they are asking for (there is a difference in how the reset gate is applied). In term of accuracy, it gives the same as the GRU from pytorch which is good sign.